zadanie domowe po 3 dniu

+++ src/main/java/pl/coderslab/controller/AuthorController.java
@@ -4,23 +4,34 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import pl.coderslab.dao.ArticleDao;
 import pl.coderslab.dao.AuthorDao;
+import pl.coderslab.entity.Article;
 import pl.coderslab.entity.Author;
 
 import javax.validation.Valid;
 import java.util.Collection;
+import java.util.List;
 
 @Controller
 public class AuthorController {
     
     private final AuthorDao authorDao;
+    private final ArticleDao articleDao;
 
-    public AuthorController(AuthorDao authorDao) {
+    public AuthorController(AuthorDao authorDao, ArticleDao articleDao) {
         this.authorDao = authorDao;
+        this.articleDao = articleDao;
     }
 
     @RequestMapping("/deleteauthor/{id}")
     public String delete(@PathVariable long id){
+
+        List<Article> list = articleDao.findAuthorById(id);
+        if (!(list.size() == 0)) {
+            return "noDelete";
+        }
+
         Author author = authorDao.findById(id);
         authorDao.delete(author);
         return "redirect:/listofauthors";
